+++ db/migrate/20221008040859_create_posts.rb
@@ -1,10 +1,11 @@
 class CreatePosts < ActiveRecord::Migration[6.1]
   def change
     create_table :posts do |t|
-      t.integer  :customer_id  , null: false
-      t.integer  :prefecture_id, null: false
+      t.bigint   :customer_id  , null: false
+      t.bigint   :prefecture_id, null: false
       t.string   :food_name    , null: false
       t.text     :introduction , null: false
+
       t.timestamps
     end
   end
