+++ db/migrate/20221008040935_create_food_comments.rb
@@ -1,10 +1,11 @@
 class CreateFoodComments < ActiveRecord::Migration[6.1]
   def change
     create_table :food_comments do |t|
-      t.integer :post_id    , null: false
-      t.integer :customer_id, null: false
+      t.bigint  :post_id    , null: false
+      t.bigint  :customer_id, null: false
       t.text    :comment    , null: false
       t.string  :star
+
       t.timestamps
     end
   end
