+++ db/migrate/20221008041436_create_favorites.rb
@@ -1,8 +1,9 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.integer  :post_id    , null: false
-      t.integer  :customer_id, null: false
+      t.bigint  :post_id    , null: false
+      t.bigint  :customer_id, null: false
+
       t.timestamps
     end
   end
