+++ db/migrate/20221012042653_create_groups.rb
@@ -1,9 +1,10 @@
 class CreateGroups < ActiveRecord::Migration[6.1]
   def change
     create_table :groups do |t|
-      t.integer :owner_customer_id, null: false
-      t.string  :name,              null: false
-      t.text    :introduction,      null: false
+      t.bigint :owner_customer_id, null: false
+      t.string :name,              null: false
+      t.text   :introduction,      null: false
+
       t.timestamps
     end
   end
