+++ db/migrate/20221012042713_create_group_customers.rb
@@ -1,8 +1,9 @@
 class CreateGroupCustomers < ActiveRecord::Migration[6.1]
   def change
     create_table :group_customers do |t|
-      t.integer :customer_id, null: false
-      t.integer :group_id,    null: false
+      t.bigint :customer_id, null: false
+      t.bigint :group_id,    null: false
+
       t.timestamps
     end
   end
